+++ include/vku/shaderModule.hpp
@@ -29,7 +29,7 @@ public:
     printf("%d\n", (int)length);
     if (length <= 0) throw(std::runtime_error("shaderModule: file not found or empty"));
 
-    std::vector<uint8_t> buf(length);
+    std::vector<uint8_t> buf((size_t)length);
     input.seekg(0);
     input.read((char*)buf.data(), buf.size());
     if (buf.size() == 0) throw(std::runtime_error("shaderModule(): shader file empty or not found"));
